\name{ISU.Stat544-package}
\alias{ISU.Stat544-package}
\alias{ISU.Stat544}
\docType{package}
\title{
Stat544: Bayesian Statistics
}
\description{
Misc functions and datasets for the course Stat544 in the Department of Statistics, Iowa State University.
}
\details{
\tabular{ll}{
Package: \tab ISU.Stat544\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-01-12\cr
License: \tab GPL-2 | GPL-3\cr
LazyLoad: \tab yes\cr
}
An overview to come here.
}
\author{
Yihui Xie

Maintainer: Yihui Xie <xie@yihui.name>
}
\references{
To come
}
\keyword{ package }
\examples{
## TODO
}
