\name{rivers}
\alias{rivers}
\docType{data}
\title{
rivers
}
\description{
rivers
}
\usage{data(rivers)}
\format{
  The format is:
 num [1:141] 735 320 325 392 524 ...
}
\details{
I don't know.
}
\source{
I don't know.
}
\references{
I don't know.
}
\examples{
data(rivers)
## do whatever you want then
}
\keyword{datasets}
