\name{square_foot}
\alias{square_foot}
\docType{data}
\title{
square_foot
}
\description{
square_foot
}
\usage{data(square_foot)}
\format{
  The format is:
 num [1:203] 612 698 873 709 892 841 854 944 708 640 ...
}
\details{
I don't know.
}
\source{
I don't know.
}
\references{
I don't know.
}
\examples{
data(square_foot)
}
\keyword{datasets}
